+++ engine/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Split ex_webrtc integration tests to JSON and protobuf [#13](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/13)
 * Send all endpoints metadata on ready notification [#14](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/14)
 * Rename `encoding` to `variant` when refering to simulcast layers [#15](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/15)
+* Use :protobuf as serializer in tests [#30](https://github.com/fishjam-cloud/membrane_rtc_engine/pull/30)
 
 ## 0.23.0
 * Add RTCP sender reports [#393](https://github.com/fishjam-dev/membrane_rtc_engine/pull/393)